% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallingCNV.R
\name{CallingCNV}
\alias{CallingCNV}
\title{Call CNV states with Gaussian mixture modeling}
\usage{
CallingCNV(object, modelNames = "E")
}
\arguments{
\item{object}{A \code{CNVisionObj} with \code{ScaledCNV} from \code{InferPloidy()}.}

\item{modelNames}{A vector of character strings indicating the models to be
fitted in the EM phase of clustering (see \code{mclust::Mclust}).}
}
\value{
A \code{CNVisionObj} with GMM results stored in \code{object@result$mclust}
and summary statistics in \code{object@result$classification}.
}
\description{
Fit a Gaussian mixture model (GMM) to scaled segment means and summarize
component statistics for CNV calling.
}
\examples{
\dontrun{
obj <- CallingCNV(obj, modelNames = "E")
}
}
