% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chrom.sizes}
\alias{chrom.sizes}
\title{hg19 chromosome sizes}
\format{
A data frame with columns \code{chrom}, \code{length}, and \code{offset}.
}
\description{
Chromosome sizes and cumulative offsets for hg19.
}
\examples{
\dontrun{
data(chrom.sizes, package = "CNVision")
}
}
