% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-generics.R
\docType{class}
\name{CNVisionObj-class}
\alias{CNVisionObj-class}
\title{The CNVisionObj Class}
\description{
Define the 'CNVision' class, containing the 'data' class and other slots
}
\section{Slots}{

\describe{
\item{\code{cnvData}}{DFrame.}

\item{\code{bin}}{GRanges.}

\item{\code{result}}{list.}

\item{\code{config}}{list.}
}}

