% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mask.R
\name{FilterBins}
\alias{FilterBins}
\title{Filter bins or rows in CNVisionObj}
\usage{
FilterBins(object, index, keep = TRUE)
}
\arguments{
\item{object}{A CNVisionObj object}

\item{index}{A logical, integer, or character vector indicating bins to keep or drop}

\item{keep}{If TRUE (default), keep the bins in index; if FALSE, remove them}
}
\value{
A filtered CNVisionObj object
}
\description{
Filter bins or rows in CNVisionObj
}
