% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19_bin}
\alias{hg19_bin}
\title{Precomputed hg19 genomic bins}
\format{
A named list of \code{GRanges} objects.
}
\description{
A list of \code{GRanges} objects keyed by resolution and read length.
Used by \code{LoadGenomeBins()}.
}
\examples{
\dontrun{
data(hg19_bin, package = "CNVision")
}
}
