% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-generics.R, R/InferPloidy.R
\name{InferPloidy}
\alias{InferPloidy}
\alias{InferPloidy,CNVisionObj-method}
\title{Infer ploidy state of the CNVision object}
\usage{
InferPloidy(object, ...)

\S4method{InferPloidy}{CNVisionObj}(
  object,
  m_start = 1.5,
  m_end = 6,
  m_step = 0.05,
  b_start = -0.25,
  b_end = 0.25,
  b_step = 0.01
)
}
\arguments{
\item{object}{A CNVisionObj}

\item{...}{Optional parameters: m_start, m_end, m_step, b_start, b_end, b_step
controlling the search grid for multiplier and offset.}

\item{m_start}{Start value for multiplier grid.}

\item{m_end}{End value for multiplier grid.}

\item{m_step}{Step size for multiplier grid.}

\item{b_start}{Start value for offset grid.}

\item{b_end}{End value for offset grid.}

\item{b_step}{Step size for offset grid.}
}
\value{
The \code{CNVisionObj} with ploidy results stored in \code{object@result},
including \code{ErrorGrid}, \code{optimal_scale}, \code{ploidy}, and \code{ScaledCNV}.
}
\description{
Estimate sample-wide ploidy using grid search on segment means
}
