% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-generics.R, R/Binning.R
\name{LoadBins}
\alias{LoadBins}
\alias{LoadBins,CNVisionObj-method}
\title{Load bins into CNVision object}
\usage{
LoadBins(object, resolution, length, genome)

\S4method{LoadBins}{CNVisionObj}(object, resolution = 600, length = 150, genome = c("hg38", "hg19"))
}
\arguments{
\item{object}{A \code{CNVisionObj} object.}

\item{resolution}{Integer. Bin resolution in base pairs. Must be one of:
\code{c(5, 10, 20, 100, 200, 400, 600, 800, 1000, 2000, 3000, 5000, 6000)}.
Default is 600.}

\item{length}{Integer. Read length used for binning. Default is 150.}

\item{genome}{Character. Reference genome. One of \code{"hg38"}, \code{"hg19"}. Default is \code{"hg19"}.}
}
\value{
A \code{CNVisionObj} with updated bin and config slot.

A \code{CNVisionObj} with updated bin and config slot.
}
\description{
Load bins into CNVision object

Load bins into CNVision object
}
