% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binning.R
\name{LoadGenomeBins}
\alias{LoadGenomeBins}
\title{Load Genome Bins}
\usage{
LoadGenomeBins(resolution = 600, length = 150, genome = c("hg38", "hg19"))
}
\arguments{
\item{resolution}{Integer. Bin resolution in base pairs. Must be one of:
\code{c(5, 10, 20, 100, 200, 400, 600, 800, 1000, 2000, 3000, 5000, 6000)}.
Default is 600.}

\item{length}{Integer. Read length used for binning. Default is 150.}

\item{genome}{Character. Reference genome. One of \code{"hg38"}, \code{"hg19"}. Default is \code{"hg19"}.}
}
\value{
A \code{GRanges} object containing the binned genome.
}
\description{
Loads precomputed genome bins (GRanges object) based on reference genome, bin resolution, and read length.
}
