% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mask.R
\name{Maskbins}
\alias{Maskbins}
\alias{Maskbins,CNVisionObj-method}
\title{Mask bins in CNVisionObj}
\usage{
Maskbins(
  object,
  mask_types = c("centromere", "clone", "contig", "heterochromatin", "scaffold",
    "short_arm", "telomere")
)

\S4method{Maskbins}{CNVisionObj}(
  object,
  mask_types = c("centromere", "clone", "contig", "heterochromatin", "scaffold",
    "short_arm", "telomere")
)
}
\arguments{
\item{object}{A CNVisionObj object containing bin and data information.}

\item{mask_types}{A character vector specifying types of regions to mask.
Allowed values include "none", "centromere", "clone", "contig",
"heterochromatin", "scaffold", "short_arm", and "telomere".
Use "none" to disable masking, or NULL to apply default mask types.}
}
\value{
The modified CNVisionObj object with specified bins masked.
}
\description{
This method masks specified bins in a CNVisionObj object based on given mask types.
Masking is used to exclude problematic genomic regions (e.g., centromeres, telomeres)
from downstream analyses.
}
\examples{
\dontrun{
obj <- Maskbins(obj, mask_types = c("centromere", "telomere"))
obj <- Maskbins(obj, mask_types = "none")
}

}
