% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{NormalizeByMeanPlus1}
\alias{NormalizeByMeanPlus1}
\alias{NormalizeData}
\title{Normalize bin counts and GC bias}
\usage{
NormalizeByMeanPlus1(x)

NormalizeData(object, method = "Normalize")
}
\arguments{
\item{x}{Numeric vector of counts.}

\item{object}{A \code{CNVisionObj}.}

\item{method}{Normalization method. Currently only \code{"Normalize"} is supported.}
}
\value{
Numeric vector of normalized counts.

A \code{CNVisionObj} with normalized \code{ratio} and GC-corrected \code{gc.ratio}.
}
\description{
Normalize bin counts and GC bias
}
\examples{
\dontrun{
obj <- NormalizeData(obj)
}
}
