% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-generics.R, R/Plot.R
\name{plotCNV}
\alias{plotCNV}
\alias{plotCNV,CNVisionObj-method}
\title{Plot CNV profile}
\usage{
plotCNV(object, cell, chr = NULL, annotate = TRUE, without_x = TRUE)

\S4method{plotCNV}{CNVisionObj}(object, cell, chr = NULL, annotate = TRUE, without_x = TRUE)
}
\arguments{
\item{object}{A CNVisionObj}

\item{cell}{Character. Target cell/sample name.}

\item{chr}{Optional vector of chromosome names (e.g. "chr1", "chrX"). Default is all.}

\item{annotate}{Logical. Add CV/NrcD annotation text.}

\item{without_x}{Logical. Hide x-axis labels (useful for stacked plots).}
}
\value{
A ggplot object.
}
\description{
Plot CNV profile
}
