% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-generics.R, R/Plot.R
\name{plotPeaks}
\alias{plotPeaks}
\alias{plotPeaks,CNVisionObj-method}
\title{Plot CNV Peaks}
\usage{
plotPeaks(object)

\S4method{plotPeaks}{CNVisionObj}(object)
}
\arguments{
\item{object}{A \code{CNVisionObj} instance containing CNV analysis results.}
}
\value{
A \code{ggplot} or \code{patchwork} object displaying CNV peaks.
}
\description{
This S4 method plots CNV peaks for a \code{CNVisionObj} object.
It visualizes both histogram of CNV segments and density/peak information.
}
\examples{
\dontrun{
  data(myCNVObj)
  plotPeaks(myCNVObj)
}

}
