% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-generics.R, R/Plot.R
\name{PlotPloidy}
\alias{PlotPloidy}
\alias{PlotPloidy,CNVisionObj-method}
\title{Plot Ploidy Estimation Grid}
\usage{
PlotPloidy(object, cell)

\S4method{PlotPloidy}{CNVisionObj}(object, cell)
}
\arguments{
\item{object}{A CNVisionObj}

\item{cell}{Character. Target cell/sample name.}
}
\value{
A ggplot2 object (and it will be printed)
}
\description{
Plot the heatmap of (multiplier, offset) vs error from ploidy estimation.
Requires that \code{InferPloidy()} has been run first.
}
